Move Session logic
- Move Session decorator & interface to common module
- Create LazySessionPipe to lazily reference real SessionPipe (indirectly) which is injected via SESSION_PIPE_TOKEN
- Have the AuthModule provide SESSION_PIPE_TOKEN via SessionPipe

This way everyone is referencing session code from the common module
and only auth module knows about the real logic.
This fixes the circular dependency problem we were having with the decorator.
