+++ src/common/index.ts
@@ -13,3 +13,4 @@ export * from './resource.dto';
 export * from './secured-list';
 export * from './secured-property';
 export * from './sensitivity.enum';
+export { ISession, Session } from './session';
