+++ src/components/auth/index.ts
@@ -1,3 +1,2 @@
-export * from './session';
 export * from './auth.service';
 export * from './auth.module';
