+++ src/components/budget/budget.service.ts
@@ -4,8 +4,8 @@ import {
   NotImplementedException,
 } from '@nestjs/common';
 import { generate } from 'shortid';
+import { ISession } from '../../common';
 import { DatabaseService, ILogger, Logger } from '../../core';
-import { ISession } from '../auth';
 import {
   Budget,
   BudgetListInput,
