+++ src/components/ceremony/ceremony.resolver.ts
@@ -1,6 +1,5 @@
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
-import { IdArg } from '../../common';
-import { ISession, Session } from '../auth';
+import { IdArg, ISession, Session } from '../../common';
 import { CeremonyService } from './ceremony.service';
 import {
   Ceremony,
