+++ src/components/project/project-member/project-member.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { generate } from 'shortid';
+import { ISession } from '../../../common';
 import { DatabaseService, ILogger, Logger } from '../../../core';
-import { ISession } from '../../auth';
 import { RedactedUser, User, UserService } from '../../user';
 import {
   CreateProjectMember,
