+++ src/components/project/project.service.ts
@@ -5,9 +5,8 @@ import {
 } from '@nestjs/common';
 import { DateTime } from 'luxon';
 import { generate } from 'shortid';
-import { Sensitivity } from '../../common';
+import { ISession, Sensitivity } from '../../common';
 import { DatabaseService, ILogger, Logger } from '../../core';
-import { ISession } from '../auth';
 import {
   EngagementListInput,
   SecuredInternshipEngagementList,
