+++ src/components/user/user.service.ts
@@ -3,8 +3,8 @@ import * as argon2 from 'argon2';
 import { node, relation } from 'cypher-query-builder';
 import { DateTime } from 'luxon';
 import { generate } from 'shortid';
+import { ISession } from '../../common';
 import { DatabaseService, ILogger, Logger, OnIndex } from '../../core';
-import { ISession } from '../auth';
 import {
   OrganizationListInput,
   OrganizationService,
